Inicio do projeto, organizei o template, fiz as camadas
