+++ Modulo-6/projeto-labook/src/app.ts
@@ -1,0 +1,10 @@
+import express, { Express, Request, Response } from "express"
+import cors from "cors"
+
+const app: Express = express()
+app.use(express.json())
+app.use(cors())
+
+app.listen(3003, () => {
+    console.log("Server running on port 3003")
+ })
