+++ Modulo-6/projeto-labook/src/data/BaseDatabase.ts
@@ -1,0 +1,22 @@
+import cors from "cors"
+import knex from "knex"
+import dotenv from "dotenv"
+import Knex from "knex"
+
+
+dotenv.config()
+
+export abstract class BaseDatabase{
+    protected static connection: Knex = knex({
+        client: "mysql",
+        connection: {
+           host: process.env.DB_HOST,
+           user: process.env.DB_USER,
+           password: process.env.DB_PASSWORD,
+           database: process.env.DB_SCHEMA,
+           port: 3306,
+           multipleStatements: true
+        }
+     })
+     
+}
