+++ Modulo-6/projeto-labook/src/index.ts
@@ -1,92 +1,22 @@
-/**************************** IMPORTS ******************************/
-
-import express, { Express, Request, Response } from "express"
-import cors from "cors"
-import knex from "knex"
-import dotenv from "dotenv"
-import { v4 } from "uuid"
-import Knex from "knex"
-
-/**************************** CONFIG ******************************/
-
-dotenv.config()
-
-export const connection: Knex = knex({
-   client: "mysql",
-   connection: {
-      host: process.env.DB_HOST,
-      user: process.env.DB_USER,
-      password: process.env.DB_PASSWORD,
-      database: process.env.DB_SCHEMA,
-      port: 3306,
-      multipleStatements: true
-   }
-})
-
-const app: Express = express()
-app.use(express.json())
-app.use(cors())
-
-/**************************** TYPES ******************************/
-
-type authenticationData = {
-   id: string
-}
-
-type user = {
-   id: string,
-   name: string,
-   email: string,
-   password: string
-}
-
-enum POST_TYPES {
-   NORMAL = "normal",
-   EVENT = "event"
-}
-
-type post = {
-   id: string,
-   photo: string,
-   description: string,
-   type: POST_TYPES,
-   createdAt: Date,
-   authorId: string
-}
-
-/**************************** SERVICES ******************************/
-
-const generateId = (): string => v4()
 
 /**************************** ENDPOINTS ******************************/
 
-app.post('/users', async (req: Request, res: Response) => {
+app.post('/users', async  => {
    try {
-      let message = "Success!"
-      const { name, email, password } = req.body
+      
 
       if (!name || !email || !password) {
          res.statusCode = 406
          message = '"name", "email" and "password" must be provided'
-         throw new Error(message)
+         
       }
 
-      const id: string = generateId()
 
-      await connection('labook_users')
-         .insert({
-            id,
-            name,
-            email,
-            password
-         })
+      
 
-      res.status(201).send({ message })
 
    } catch (error:any) {
-      res.statusCode = 400
-      let message = error.sqlMessage || error.message
-      res.send({ message })
+      
    }
 })
 
@@ -152,6 +82,3 @@ app.get('/posts/:id', async (req: Request, res: Response) => {
 
 /**************************** SERVER INIT ******************************/
 
-app.listen(3003, () => {
-   console.log("Server running on port 3003")
-})
