+++ Modulo-6/projeto-labook/src/model/POST_TYPES.ts
@@ -1,0 +1,4 @@
+enum POST_TYPES {
+    NORMAL = "normal",
+    EVENT = "event"
+ }
