+++ Modulo-6/projeto-labook/src/model/Post.ts
@@ -1,0 +1,8 @@
+type post = {
+    id: string,
+    photo: string,
+    description: string,
+    type: POST_TYPES,
+    createdAt: Date,
+    authorId: string
+ }
