+++ Modulo-6/projeto-labook/src/model/UserDTO.ts
@@ -1,0 +1,6 @@
+type user = {
+    id: string,
+    name: string,
+    email: string,
+    password: string
+ }
