+++ Modulo-6/projeto-labook/src/model/authenticationData.ts
@@ -1,0 +1,3 @@
+type authenticationData = {
+    id: string
+ }
