+++ Modulo-6/projeto-labook/src/services/generateId.ts
@@ -1,0 +1,3 @@
+import { v4 } from "uuid"
+
+export const generateId = (): string => v4()
